+++ lib/injexor.ex
@@ -11,21 +11,21 @@ defmodule Injexor do
   It can be either a single module or a list.
 
   ```
-  use Injexor, otp_app: :my_app, inject: MyApp.Repo
-  use Injexor, otp_app: :my_app, inject: [MyApp.Repo, MyApp.Repo]
+  use Injexor, inject: MyApp.Repo
+  use Injexor, inject: [MyApp.Repo, MyApp.Repo]
   ```
 
   You can also explicitly tell it what behaviour to use in the case where a module may have many.
 
   ```
-  use Injexor, otp_app: :my_app, inject: {MyApp.Repo, MyApp.Repo.Behaviour}
+  use Injexor, inject: {MyApp.Repo, MyApp.Repo.Behaviour}
   ```
 
   The other way is to define `@inject` attribute above a function def. `@inject` also accepts either
   a single module or a list. This should only be used when you want to affect a single function.
 
   ```
-  use Injexor, otp_app: :my_app
+  use Injexor
 
   alias MyApp.Repo
 
@@ -37,16 +37,26 @@ defmodule Injexor do
 
   Then you can call the module as normal in your code and use mox/hammox in your tests.
 
-  `otp_app` is used to look up the module to inject in.
+  For example for `use Injexor, inject: MyApp.Repo`, the macro will lookup the module
+  to inject with `Application.get_env(:injexor, MyApp.Repo)[:inject]`.
 
-  For example for `use Injexor, otp_app: :my_app, inject: MyApp.Repo`, the macro will lookup the module
-  to inject with `Application.get_env(:my_app, MyApp.Repo)`. In testing this will default to
-  `MyApp.Team.Mock` if not specified in config, otherwise the default is the module itself.
+  Example config
+
+  ```
+  config :injexor, MyApp.Repo, inject: MyApp.Repo.Mock
+  ```
+
+  You can altneratively setup a default that will be appended to all modules.
+
+  ```
+  config :injexor, :default, Mock
+  ```
+
+  If none of the above are set, the module itself will be used.
   """
 
   defmacro __using__(opts) do
     Module.put_attribute(__CALLER__.module, :inject_all, opts[:inject])
-    Module.put_attribute(__CALLER__.module, :otp_app, opts[:otp_app])
 
     quote do
       @on_definition Injexor.Hooks