+++ lib/injexor/hooks.ex
@@ -1,4 +1,6 @@
 defmodule Injexor.Hooks do
+  @moduledoc false
+
   def __on_definition__(env, kind, name, args, guards, body) do
     definitions = Module.get_attribute(env.module, :definitions) || []
     arity = length(args)
@@ -57,8 +59,6 @@ defmodule Injexor.Hooks do
       |> Enum.map(fn {alias_as, alias_module} -> {to_string(alias_as), alias_module} end)
       |> Map.new()
 
-    otp_app = Module.get_attribute(env.module, :otp_app)
-
     definitions =
       Module.get_attribute(env.module, :definitions)
       # we need to reverse as the definitions were appended to the module attribute
@@ -68,7 +68,7 @@ defmodule Injexor.Hooks do
         {kind, name, arity}
       end)
       |> Enum.map(fn {{kind, name, arity}, definitions} ->
-        definitions = Enum.map(definitions, &process_definition(&1, otp_app, aliases))
+        definitions = Enum.map(definitions, &process_definition(&1, aliases))
 
         {kind, name, arity, definitions}
       end)
@@ -125,7 +125,6 @@ defmodule Injexor.Hooks do
            body: body,
            injects: injects
          },
-         otp_app,
          aliases
        ) do
     call =
@@ -143,11 +142,11 @@ defmodule Injexor.Hooks do
            {{:., _line0, [{:__aliases__, _counter, _module_parts}, _function]}, _line1, _args} =
                ast
          ]} ->
-          inject_ast = inject_ast(ast, injects, otp_app, aliases, true)
+          inject_ast = inject_ast(ast, injects, aliases, true)
           {:|>, pipe_line, [piped_ast, inject_ast]}
 
         {{:., _line0, [{:__aliases__, _counter, _module_parts}, _function]}, _line1, _args} = ast ->
-          inject_ast(ast, injects, otp_app, aliases, false)
+          inject_ast(ast, injects, aliases, false)
 
         other ->
           other
@@ -156,8 +155,8 @@ defmodule Injexor.Hooks do
     {call, body}
   end
 
-  defp inject_ast(ast, injects, otp_app, aliases, is_piping) do
-    {{:., line0, [{:__aliases__, counter, module_parts}, function]}, line1, args} = ast
+  defp inject_ast(ast, injects, aliases, is_piping) do
+    {{:., line0, [{:__aliases__, _counter, module_parts}, function]}, line1, args} = ast
     arity = if is_piping, do: length(args) + 1, else: length(args)
     module = Module.concat(module_parts)
     full_module = module_from_alias(aliases, module)
@@ -167,19 +166,29 @@ defmodule Injexor.Hooks do
            Enum.find(injects, fn {inject, _behaviour} ->
              full_module == inject or module == inject
            end),
-         inject_module_parts <- inject_module(otp_app, full_module, behaviour),
+         inject_module <- inject_module(full_module, behaviour),
          true <- {function, arity} in behaviour.behaviour_info(:callbacks) do
-      {{:., line0, [{:__aliases__, counter, inject_module_parts}, function]}, line1, args}
+      {{:., line0, [inject_module, function]}, line1, args}
     else
       _false ->
         ast
     end
   end
 
-  defp inject_module(otp_app, module, behaviour) do
-    Application.get_env(otp_app, behaviour, module)
-    |> Module.split()
-    |> Enum.map(&String.to_atom/1)
+  defp inject_module(module, behaviour) do
+    default =
+      case Application.get_env(:injexor, :default) do
+        nil -> module
+        default -> Module.concat(module, default)
+      end
+
+    quote do
+      case Application.get_env(:injexor, unquote(behaviour)) do
+        opts when is_list(opts) -> opts[:inject] || unquote(default)
+        nil -> unquote(default)
+        module -> module
+      end
+    end
   end
 
   defp behaviour(module) do
@@ -200,7 +209,6 @@ defmodule Injexor.Hooks do
 
         ```
         use Injexor,
-          otp_app: :my_app,
           inject: [{Inject.Example, Inject.ExampleBehaviour}]
         ```
         """
@@ -213,7 +221,6 @@ defmodule Injexor.Hooks do
 
         ```
         use Injexor,
-          otp_app: :my_app,
           inject: [{Inject.Example, Inject.ExampleBehaviour}]
         ```
         """