+++ test/support/another_behaviour.ex
@@ -1,0 +1,3 @@
+defmodule InjexorTest.Inject.AnotherBehaviour do
+  @callback call(any()) :: any()
+end