+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+1.0.0
+------
+#### Changes
+- Now evaluates modules to inject at runtime
+- Supports passing :default option that will be appended to modules to automatically inject to make test config simpler
+
 0.1.0
 ------
 #### Changes